Se agregaron archivos de registro para la ejecución, evaluación y persona de microservicios; se incluyeron copias de seguridad del esquema y un script de restablecimiento de la base de datos.
- Se crearon archivos de registro para microservicio_ejecucion, microservicio_evaluacion y microservicio_persona con registros de inicio detallados.
- Se agregó schema.sql.bak para el microservicio persona, lo que permite definir la estructura de la base de datos.
- Se implementó el script resetear-bd.bat para automatizar el proceso de restablecimiento de la base de datos, incluyendo comprobaciones del servicio MySQL y solicitudes de confirmación.

+++ microservicio_persona/src/main/java/com/edutech/microservicio_persona/controller/PersonaController.java
@@ -42,7 +42,7 @@ public class PersonaController {
     }
 
     @GetMapping("/tipo/{idTipoPersona}")
-    public ResponseEntity<List<Persona>> getPersonasByTipo(@PathVariable Integer idTipoPersona) {
+    public ResponseEntity<List<Persona>> getPersonasByTipo(@PathVariable Long idTipoPersona) {
         List<Persona> personas = personaService.findByTipoPersona(idTipoPersona);
         return new ResponseEntity<>(personas, HttpStatus.OK);
     }
