+++ microservicio_persona/src/main/java/com/edutech/microservicio_persona/repository/TipoPersonaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoPersonaRepository extends JpaRepository<TipoPersona, Integer> {
+public interface TipoPersonaRepository extends JpaRepository<TipoPersona, Long> {
 }
